chatgpt-app/requirements.txt
PySide6
openai
asyncio
qasync
qt-material
python-dotenv
pyinstaller